requirements.txt
fastapi
fastapi[standard]
uvicorn
streamlit
groq
faiss-cpu
PyPDF2
sentence-transformers
pydantic
python-multipart
numpy
